refactor: 비밀번호 재설정 페이지 코드 리팩토링 (#10)
* feat: react-hook-form 라이브러리 설치

* feat: yup 라이브러리 설치

* feat: 비밀번호 표시/숨기기 아이콘, sns 로고 아이콘 추가

* feat: FormField 컴포넌트(입력 필드, 비밀번호 표시/숨기기 아이콘, 에러 메시지) 추가

* feat: 회원가입, 로그인 스키마 추가

* feat: @hookform/resolvers, yup 설치

* feat: 회원가입 기본 디자인 구현

* chore: assets, styles 폴더 > root 폴더로 이동

* fix: svg 파일 import 경로 수정

* fix: rgba 형식으로 border 색상 값 수정

* feat: 테마 확장(screens, width, gap, padding, margin)

* feat: FormField 컴포넌트에 Tailwind CSS 적용

* feat: signup 페이지에 Tailwind CSS 적용

* chore: src 폴더로 app, assets, components, styles, utils 폴더 이동

* style: lg 사이즈 반응형 css 추가

* feat: 로그인 페이지 기본 디자인 구현

* feat: 테마 확장(gap, margin)

* feat: globals.css에 유틸리티 클래스 추가

* feat: login 페이지에 Tailwind CSS 적용 및 반응형 구현

* feat: 로그인 버튼 유효성 검사 적용

* chore: 스타일 파일 경로 절대 경로로 변경

* feat: 비밀번호 재설정 스키마 추가

* feat: 비밀번호 재설정 페이지 기본 디자인 구현

* fix: width에 '9.5-custom' 추가, gap 중복값 제거 in tailwind.config.js

* feat: 비밀번호 재설정 페이지 Tailwind CSS 적용 및 반응형 구현

* fix: FormField 컴포넌트 모바일 사이즈 폰트 크기 수정

* feat: 커스텀 배경 그라디언트 추가 in tailwind.config.js

* feat: FormField 컴포넌트 모바일 사이즈 linear gradient 추가

* feat: 비밀번호 재설정 페이지 퍼블리싱 (#5)

* chore: 스타일 파일 경로 절대 경로로 변경

* feat: 비밀번호 재설정 스키마 추가

* feat: 비밀번호 재설정 페이지 기본 디자인 구현

* fix: width에 '9.5-custom' 추가, gap 중복값 제거 in tailwind.config.js

* feat: 비밀번호 재설정 페이지 Tailwind CSS 적용 및 반응형 구현

* fix: FormField 컴포넌트 모바일 사이즈 폰트 크기 수정

* feat: 커스텀 배경 그라디언트 추가 in tailwind.config.js

* feat: FormField 컴포넌트 모바일 사이즈 linear gradient 추가

* feat: 팀 페이지 퍼블리싱 (#6)

* docs: 타입 폴더 생성

* feat: 팀페이지 팀 설정 및 할 일 목록 퍼블리싱

* chore: 원형 프로그레스바 설치

* feat: 리포트 퍼블리싱(반응형 미완성)

* feat: 모바일 반응형 수정

* feat: 멤버 리스트 퍼블리싱 및 랜덤 컬러칩 생성

* docs: 파읾명 공백 제거

* css: 텍스트 그라데이션 적용

* feat: 완료 task 아이콘 추가

* refactor: 회원가입 페이지 코드 리팩토링 (#7)

* feat: react-hook-form 라이브러리 설치

* feat: yup 라이브러리 설치

* feat: 비밀번호 표시/숨기기 아이콘, sns 로고 아이콘 추가

* feat: FormField 컴포넌트(입력 필드, 비밀번호 표시/숨기기 아이콘, 에러 메시지) 추가

* feat: 회원가입, 로그인 스키마 추가

* feat: @hookform/resolvers, yup 설치

* feat: 회원가입 기본 디자인 구현

* chore: assets, styles 폴더 > root 폴더로 이동

* fix: svg 파일 import 경로 수정

* fix: rgba 형식으로 border 색상 값 수정

* feat: 테마 확장(screens, width, gap, padding, margin)

* feat: FormField 컴포넌트에 Tailwind CSS 적용

* feat: signup 페이지에 Tailwind CSS 적용

* chore: src 폴더로 app, assets, components, styles, utils 폴더 이동

* style: lg 사이즈 반응형 css 추가

* feat: 회원가입 스키마 수정

* feat: 회원가입 버튼 유효성 검사 적용

* chore: css 파일 절대경로로 수정

* feat: 공통, 회원가입, 로그인 타입 분리

* refactor: 타입 삭제

* refactor: 회원가입 스키마에 image 필드 추가, 비밀번호 확인 필드명 변경

* feat: 비밀번호 표시 상태 관리를 위한 Zustand store 추가

* feat: 회원가입 필드 데이터 커스텀 훅으로 분리

* refactor: 타입 및 불필요한 코드 삭제, 컴포넌트명+비밀번호 확인 변수명+state명 변경

---------


* refactor: 타입 삭제, 컴포넌트명 변경

* feat: 로그인 필드 데이터 useLoginFieldData 훅 추가

* refactor: 로그인 필드 데이터 반복 코드 useLoginFieldData로 리팩토링, 불필요한 코드 제거

* refactor: components/signup 폴더명 components/auth로 변경

* refactor: 비밀번호 재설정 스키마 필드명 및 정규식 수정, token 필드 추가

* refactor: 비밀번호 재설정 타입 분리

* refactor: 타입명, 비밀번호 변수명 및 컴포넌트명 변경

* fix: 비밀번호 재설정 페이지 placeholder 수정

* fix: 스키마 수정

* refactor: 필드 데이터 반복 코드 useResetPasswordFieldData로 리팩토링, 불필요한 코드 제거

* fix: 비밀번호 검증 로직 수정

* refactor: tailwind.config.js 중복 유틸리티 클래스(custom-gradient) 제거

---------
